tbd-encoder/encoder_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
# TensorFlow 2.15 is the target version for the rebuild logic
tensorflow-cpu==2.15.0
h5py==3.10.0
numpy>=1.26.0,<2.0.0